issues-383 Remove postgres example
