--- examples/postgres_json/Cargo.toml
@@ -1,25 +1,0 @@
-[workspace]
-# A separate workspace
-
-[package]
-name = "sea-query-postgres-json-example"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-chrono = { version = "^0", default-features = false, features = ["clock"] }
-time = { version = "^0.3", features = ["macros"] }
-uuid = { version = "^1", features = ["serde", "v4"] }
-serde_json = "^1"
-rust_decimal = { version = "^1" }
-postgres = "^0.19"
-sea-query = { path = "../../", features = [
-    "postgres",
-    "postgres-rust_decimal",
-    "postgres-array",
-    "with-uuid",
-    "with-chrono",
-    "with-json",
-    "with-time",
-] }
-# NOTE: if you are copying this example into your own project, replace `path = ` with `version = "^0"`
